Fix backend/api requirements and .gitignore hiding frontend/lib/
Two fixes:
- .gitignore's Python `lib/` rule (for venvs) had no leading slash, so it
  matched frontend/lib/ at any depth too. api.ts, types.ts, store.ts, and
  utils.ts were never committed to this repo, in any prior commit -- the
  actual GitHub repo has been missing every @/lib import target this whole
  time. (The "Fix Render alias resolution" / tsconfig-paths commits in this
  repo's history look like symptoms of this: patching the alias while the
  files behind it were silently absent.) Anchored the rule to /lib/ so it
  only excludes a root-level venv dir, and added the 4 files.
- Also fixes the JSON-serialization of the rule-based affect fallback
  (numpy.float32 -> float), which is now load-bearing for the trimmed
  torch-free Vercel deployment below.

Deploy backend on Vercel (free tier) alongside the frontend:
- backend/api/index.py: ASGI entrypoint importing the existing FastAPI app
  unchanged.
- backend/api/requirements.txt: trimmed deps (drop torch + unused
  scikit-learn to fit the function size limit); the affect estimator's
  existing rule-based fallback covers live inference. /ml/model-card now
  reports which inference path (mlp vs rule_based) is actually live.
- backend/vercel.json: route everything to the function.
- New GET /crew/{id}/bio-live (same consent enforcement as the WS route).

Replace WebSocket streams with polling, since Vercel's Python runtime
doesn't support long-lived WebSocket connections:
- /ws/env -> poll /habitat/zones from ground/layout.tsx (also fixes zones
  not loading if a /ground/* page other than the dashboard loads first).
- /ws/bio/{id} x12 on the ground dashboard -> removed; /crew already
  returns live bio for everyone via the existing 3s poll.
- /ws/bio/{id} on the crew's own page -> poll the new bio-live endpoint.

README: document the Vercel-for-both-services deployment path and its
explicit tradeoffs (polling vs push, ephemeral state, rule-based fallback).

+++ backend/ml/affect/regressor.py
@@ -110,13 +110,15 @@ def rule_based_affect(features: np.ndarray) -> dict:
     valence = 0.3 + 0.8 * hrv_norm - 0.7 * sleep_norm - 0.3 * eda_norm + 0.2 * phase_sin
     valence = max(-1.0, min(1.0, valence))
 
-    # Feature attributions (simplified SHAP)
+    # Feature attributions (simplified SHAP). Cast to native float — these
+    # are numpy.float32 scalars (from the float32 feature vector) and are
+    # not JSON-serializable by FastAPI's default encoder otherwise.
     top_features = {
-        "hrv_rmssd": round(0.8 * hrv_norm - 0.4, 3),
-        "eda": round(-0.8 * eda_norm + 0.2, 3),
-        "sleep_debt": round(-0.7 * sleep_norm, 3),
-        "heart_rate": round(1.0 * hr_norm - 0.5, 3),
-        "lf_hf_ratio": round(0.4 * lf_hf - 0.2, 3),
+        "hrv_rmssd": round(float(0.8 * hrv_norm - 0.4), 3),
+        "eda": round(float(-0.8 * eda_norm + 0.2), 3),
+        "sleep_debt": round(float(-0.7 * sleep_norm), 3),
+        "heart_rate": round(float(1.0 * hr_norm - 0.5), 3),
+        "lf_hf_ratio": round(float(0.4 * lf_hf - 0.2), 3),
     }
 
     return {
